protect route by roles permission and alos blade file

+++ app/Http/Controllers/Backend/AdminsController.php
@@ -7,12 +7,20 @@ use App\Models\Admin;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 
 class AdminsController extends Controller
 {
     public $user;
 
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::guard('admin')->user();
+            return $next($request);
+        });
+    }
     /**
      * Display a listing of the resource.
      *
@@ -20,7 +28,9 @@ class AdminsController extends Controller
      */
     public function index()
     {
-        
+        if (is_null($this->user) || !$this->user->can('admin.view')) {
+            abort(403, 'Sorry !! You are Unauthorized to view any admin !');
+        }
 
         $admins = Admin::all();
         return view('backend.pages.admins.index', compact('admins'));
@@ -34,6 +44,10 @@ class AdminsController extends Controller
     public function create()
     {
        
+        if (is_null($this->user) || !$this->user->can('admin.create')) {
+            abort(403, 'Sorry !! You are Unauthorized to create any admin !');
+        }
+
 
         $roles  = Role::all();
         return view('backend.pages.admins.create', compact('roles'));
@@ -47,6 +61,10 @@ class AdminsController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) || !$this->user->can('admin.create')) {
+            abort(403, 'Sorry !! You are Unauthorized to create any admin !');
+        }
+
     
        // Validation Data
        $request->validate([
@@ -92,7 +110,9 @@ class AdminsController extends Controller
     public function edit($id)
     {
         
-
+        if (is_null($this->user) || !$this->user->can('admin.edit')) {
+            abort(403, 'Sorry !! You are Unauthorized to edit any admin !');
+        }
         $admin = Admin::find($id);
         $roles  = Role::all();
         return view('backend.pages.admins.edit', compact('admin', 'roles'));
@@ -107,7 +127,10 @@ class AdminsController extends Controller
      */
     public function update(Request $request, $id)
     {
-        
+        if (is_null($this->user) || !$this->user->can('admin.edit')) {
+            abort(403, 'Sorry !! You are Unauthorized to edit any admin !');
+        }
+
 
         // Create New Admin
         $admin = Admin::find($id);
@@ -146,7 +169,9 @@ class AdminsController extends Controller
     public function destroy($id)
     {
         
-
+        if (is_null($this->user) || !$this->user->can('admin.delete')) {
+            abort(403, 'Sorry !! You are Unauthorized to delete any admin !');
+        }
         $admin = Admin::find($id);
         if (!is_null($admin)) {
             $admin->delete();
