+++ app/Http/Controllers/Backend/RolesController.php
@@ -6,11 +6,20 @@ use App\User;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Permission;
 
 class RolesController extends Controller
 {
     public $user;
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::guard('admin')->user();
+            return $next($request);
+        });
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -18,6 +27,9 @@ class RolesController extends Controller
      */
     public function index()
     {
+        if (is_null($this->user) || !$this->user->can('role.view')) {
+            abort(403, 'Sorry !! You are Unauthorized to view any role !');
+        }
     
         $roles = Role::all();
         return view('backend.pages.roles.index', compact('roles'));
@@ -31,6 +43,9 @@ class RolesController extends Controller
     public function create()
     {
     
+        if (is_null($this->user) || !$this->user->can('role.create')) {
+            abort(403, 'Sorry !! You are Unauthorized to create any role !');
+        }
         $all_permissions  = Permission::all();
         $permission_groups = User::getpermissionGroups();
         return view('backend.pages.roles.create', compact('all_permissions', 'permission_groups'));
@@ -45,6 +60,9 @@ class RolesController extends Controller
     public function store(Request $request)
     {
     
+        if (is_null($this->user) || !$this->user->can('role.create')) {
+            abort(403, 'Sorry !! You are Unauthorized to create any role !');
+        }
 
         // Validation Data
         $request->validate([
@@ -86,6 +104,9 @@ class RolesController extends Controller
      */
     public function edit($id)
     {
+        if (is_null($this->user) || !$this->user->can('role.edit')) {
+            abort(403, 'Sorry !! You are Unauthorized to edit any role !');
+        }
         $role = Role::findById($id, 'admin');
         $all_permissions = Permission::all();
         $permission_groups = User::getpermissionGroups();
@@ -102,6 +123,9 @@ class RolesController extends Controller
     public function update(Request $request, $id)
     {
         
+        if (is_null($this->user) || !$this->user->can('role.edit')) {
+            abort(403, 'Sorry !! You are Unauthorized to edit any role !');
+        }
 
         // Validation Data
         $request->validate([
@@ -131,6 +155,9 @@ class RolesController extends Controller
      */
     public function destroy($id)
     {
+        if (is_null($this->user) || !$this->user->can('role.delete')) {
+            abort(403, 'Sorry !! You are Unauthorized to delete any role !');
+        }
         $role = Role::findById($id, 'admin');
         if (!is_null($role)) {
             $role->delete();
